// src/generated/types.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] }
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> }
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> }
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: string
	String: string
	Boolean: boolean
	Int: number
	Float: number
	/** Blockchain data scalar type */
	BlockchainData: any
	/** Broadcast scalar id type */
	BroadcastId: any
	/** ChainId custom scalar type */
	ChainId: any
	/** collect module data scalar type */
	CollectModuleData: any
	/** Contract address custom scalar type */
	ContractAddress: any
	/** create handle custom scalar type */
	CreateHandle: any
	/** Cursor custom scalar type */
	Cursor: any
	/** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
	DateTime: any
	/** Ens custom scalar type */
	Ens: any
	/** Ethereum address custom scalar type */
	EthereumAddress: any
	/** follow module data scalar type */
	FollowModuleData: any
	/** handle custom scalar type */
	Handle: any
	/** handle claim id custom scalar type */
	HandleClaimIdScalar: any
	/** Internal publication id custom scalar type */
	InternalPublicationId: any
	/** jwt custom scalar type */
	Jwt: any
	/** limit custom scalar type */
	LimitScalar: any
	/** Locale scalar type */
	Locale: any
	/** Markdown scalar type */
	Markdown: any
	/** mimetype custom scalar type */
	MimeType: any
	/** Nft ownership id type */
	NftOwnershipId: any
	/** Nonce custom scalar type */
	Nonce: any
	/** The notification id */
	NotificationId: any
	/** ProfileId custom scalar type */
	ProfileId: any
	/** Publication id custom scalar type */
	PublicationId: any
	/** The publication tag */
	PublicationTag: any
	/** Publication url scalar type */
	PublicationUrl: any
	/** reference module data scalar type */
	ReferenceModuleData: any
	/** Query search */
	Search: any
	/** Relayer signature */
	Signature: any
	/** Sources custom scalar type */
	Sources: any
	/** timestamp date custom scalar type */
	TimestampScalar: any
	/** The tx hash */
	TxHash: any
	/** The tx id */
	TxId: any
	/** UnixTimestamp custom scalar type */
	UnixTimestamp: any
	/** Url scalar type */
	Url: any
	/** Represents NULL values */
	Void: any
}
// ...
export type Attribute = {
	__typename?: 'Attribute'
	/** The display type */
	displayType?: Maybe<Scalars['String']>
	/** identifier of this attribute, we will update by this id  */
	key: Scalars['String']
	/** The trait type - can be anything its the name it will render so include spaces */
	traitType?: Maybe<Scalars['String']>
	/** Value attribute */
	value: Scalars['String']
}
// ...
export type CollectModule =
	| FeeCollectModuleSettings
	| FreeCollectModuleSettings
	| LimitedFeeCollectModuleSettings
	| LimitedTimedFeeCollectModuleSettings
	| RevertCollectModuleSettings
	| TimedFeeCollectModuleSettings
// ...
export type Dispatcher = {
	__typename?: 'Dispatcher'
	/** The dispatcher address */
	address: Scalars['EthereumAddress']
	/** If the dispatcher can use the relay */
	canUseRelay: Scalars['Boolean']
}
// ...
export type FeeCollectModuleSettings = {
	__typename?: 'FeeCollectModuleSettings'
	/** The collect module amount info */
	amount: ModuleFeeAmount
	contractAddress: Scalars['ContractAddress']
	/** Follower only */
	followerOnly: Scalars['Boolean']
	/** The collect module recipient address */
	recipient: Scalars['EthereumAddress']
	/** The collect module referral fee */
	referralFee: Scalars['Float']
	/** The collect modules enum */
	type: CollectModules
}
// ...
export type FollowModule = FeeFollowModuleSettings | ProfileFollowModuleSettings | RevertFollowModuleSettings
// ...
export type FollowOnlyReferenceModuleSettings = {
	__typename?: 'FollowOnlyReferenceModuleSettings'
	contractAddress: Scalars['ContractAddress']
	/** The reference modules enum */
	type: ReferenceModules
}
// ...
export type FreeCollectModuleSettings = {
	__typename?: 'FreeCollectModuleSettings'
	contractAddress: Scalars['ContractAddress']
	/** Follower only */
	followerOnly: Scalars['Boolean']
	/** The collect modules enum */
	type: CollectModules
}
// ...
export type LimitedFeeCollectModuleSettings = {
	__typename?: 'LimitedFeeCollectModuleSettings'
	/** The collect module amount info */
	amount: ModuleFeeAmount
	/** The collect module limit */
	collectLimit: Scalars['String']
	contractAddress: Scalars['ContractAddress']
	/** Follower only */
	followerOnly: Scalars['Boolean']
	/** The collect module recipient address */
	recipient: Scalars['EthereumAddress']
	/** The collect module referral fee */
	referralFee: Scalars['Float']
	/** The collect modules enum */
	type: CollectModules
}
// ...
export type LimitedTimedFeeCollectModuleSettings = {
	__typename?: 'LimitedTimedFeeCollectModuleSettings'
	/** The collect module amount info */
	amount: ModuleFeeAmount
	/** The collect module limit */
	collectLimit: Scalars['String']
	contractAddress: Scalars['ContractAddress']
	/** The collect module end timestamp */
	endTimestamp: Scalars['DateTime']
	/** Follower only */
	followerOnly: Scalars['Boolean']
	/** The collect module recipient address */
	recipient: Scalars['EthereumAddress']
	/** The collect module referral fee */
	referralFee: Scalars['Float']
	/** The collect modules enum */
	type: CollectModules
}
// ...
export type MediaSet = {
	__typename?: 'MediaSet'
	/** Medium media - will always be null on the public API */
	medium?: Maybe<Media>
	/** Original media */
	original: Media
	/** Small media - will always be null on the public API */
	small?: Maybe<Media>
}

export type MentionPublication = Comment | Post

/** The metadata attribute output */
export type MetadataAttributeOutput = {
	__typename?: 'MetadataAttributeOutput'
	/** The display type */
	displayType?: Maybe<PublicationMetadataDisplayTypes>
	/** The trait type - can be anything its the name it will render so include spaces */
	traitType?: Maybe<Scalars['String']>
	/** The value */
	value?: Maybe<Scalars['String']>
}

/** The metadata output */
export type MetadataOutput = {
	__typename?: 'MetadataOutput'
	/** The attributes */
	attributes: Array<MetadataAttributeOutput>
	/** This is the metadata content for the publication, should be markdown */
	content?: Maybe<Scalars['Markdown']>
	/** The content warning for the publication */
	contentWarning?: Maybe<PublicationContentWarning>
	/** The image cover for video/music publications */
	cover?: Maybe<MediaSet>
	/** This is the metadata description */
	description?: Maybe<Scalars['Markdown']>
	/** This is the image attached to the metadata and the property used to show the NFT! */
	image?: Maybe<Scalars['Url']>
	/** The locale of the publication,  */
	locale?: Maybe<Scalars['Locale']>
	/** The main focus of the publication */
	mainContentFocus: PublicationMainFocus
	/** The images/audios/videos for the publication */
	media: Array<MediaSet>
	/** The metadata name */
	name?: Maybe<Scalars['String']>
	/** The tags for the publication */
	tags: Array<Scalars['String']>
}
// ...
export type OnChainIdentity = {
	__typename?: 'OnChainIdentity'
	/** The ens information */
	ens?: Maybe<EnsOnChainIdentity>
	/** The POH status */
	proofOfHumanity: Scalars['Boolean']
	/** The sybil dot org information */
	sybilDotOrg: SybilDotOrgIdentity
	/** The worldcoin identity */
	worldcoin: WorldcoinIdentity
}
// ...
export type Post = {
	__typename?: 'Post'
	/** ID of the source */
	appId?: Maybe<Scalars['Sources']>
	/** The collect module */
	collectModule: CollectModule
	/** The contract address for the collect nft.. if its null it means nobody collected yet as it lazy deployed */
	collectNftAddress?: Maybe<Scalars['ContractAddress']>
	/** Who collected it, this is used for timeline results and like this for better caching for the client */
	collectedBy?: Maybe<Wallet>
	/** The date the post was created on */
	createdAt: Scalars['DateTime']
	hasCollectedByMe: Scalars['Boolean']
	/** If the publication has been hidden if it has then the content and media is not available */
	hidden: Scalars['Boolean']
	/** The internal publication id */
	id: Scalars['InternalPublicationId']
	/** The metadata for the post */
	metadata: MetadataOutput
	mirrors: Array<Scalars['InternalPublicationId']>
	/** The on chain content uri could be `ipfs://` or `https` */
	onChainContentURI: Scalars['String']
	/** The profile ref */
	profile: Profile
	reaction?: Maybe<ReactionTypes>
	/** The reference module */
	referenceModule?: Maybe<ReferenceModule>
	/** The publication stats */
	stats: PublicationStats
}

/** The social post */
export type PostMirrorsArgs = {
	by?: InputMaybe<Scalars['ProfileId']>
}

/** The social post */
export type PostReactionArgs = {
	request?: InputMaybe<ReactionFieldResolverRequest>
}

/** The Profile */
export type Profile = {
	__typename?: 'Profile'
	/** Optionals param to add extra attributes on the metadata */
	attributes?: Maybe<Array<Attribute>>
	/** Bio of the profile */
	bio?: Maybe<Scalars['String']>
	/** The cover picture for the profile */
	coverPicture?: Maybe<ProfileMedia>
	/** The dispatcher */
	dispatcher?: Maybe<Dispatcher>
	/** The follow module */
	followModule?: Maybe<FollowModule>
	/** Follow nft address */
	followNftAddress?: Maybe<Scalars['ContractAddress']>
	/** The profile handle */
	handle: Scalars['Handle']
	/** The profile id */
	id: Scalars['ProfileId']
	/** Is the profile default */
	isDefault: Scalars['Boolean']
	isFollowedByMe: Scalars['Boolean']
	isFollowing: Scalars['Boolean']
	/** Metadata url */
	metadata?: Maybe<Scalars['Url']>
	/** Name of the profile */
	name?: Maybe<Scalars['String']>
	/** The on chain identity */
	onChainIdentity: OnChainIdentity
	/** Who owns the profile */
	ownedBy: Scalars['EthereumAddress']
	/** The picture for the profile */
	picture?: Maybe<ProfileMedia>
	/** Profile stats */
	stats: ProfileStats
}
// ...
export type ProfileMedia = MediaSet | NftImage
// ...
export type ProfileStats = {
	__typename?: 'ProfileStats'
	commentsTotal: Scalars['Int']
	id: Scalars['ProfileId']
	mirrorsTotal: Scalars['Int']
	postsTotal: Scalars['Int']
	publicationsTotal: Scalars['Int']
	/** Total collects count */
	totalCollects: Scalars['Int']
	/** Total comment count */
	totalComments: Scalars['Int']
	/** Total follower count */
	totalFollowers: Scalars['Int']
	/** Total following count (remember the wallet follows not profile so will be same for every profile they own) */
	totalFollowing: Scalars['Int']
	/** Total mirror count */
	totalMirrors: Scalars['Int']
	/** Total post count */
	totalPosts: Scalars['Int']
	/** Total publication count */
	totalPublications: Scalars['Int']
}
// ...
export enum PublicationContentWarning {
	Nsfw = 'NSFW',
	Sensitive = 'SENSITIVE',
	Spoiler = 'SPOILER',
}

export type PublicationForSale = Comment | Post

/** The publication main focus */
export enum PublicationMainFocus {
	Article = 'ARTICLE',
	Audio = 'AUDIO',
	Embed = 'EMBED',
	Image = 'IMAGE',
	Link = 'LINK',
	TextOnly = 'TEXT_ONLY',
	Video = 'VIDEO',
}
// ...
export type PublicationStats = {
	__typename?: 'PublicationStats'
	commentsTotal: Scalars['Int']
	/** The publication id */
	id: Scalars['InternalPublicationId']
	/** The total amount of collects */
	totalAmountOfCollects: Scalars['Int']
	/** The total amount of comments */
	totalAmountOfComments: Scalars['Int']
	/** The total amount of mirrors */
	totalAmountOfMirrors: Scalars['Int']
	/** The total amount of upvotes */
	totalDownvotes: Scalars['Int']
	/** The total amount of downvotes */
	totalUpvotes: Scalars['Int']
}
// ...
export enum ReactionTypes {
	Downvote = 'DOWNVOTE',
	Upvote = 'UPVOTE',
}

export type ReferenceModule = FollowOnlyReferenceModuleSettings
// ...
export type RevertCollectModuleSettings = {
	__typename?: 'RevertCollectModuleSettings'
	contractAddress: Scalars['ContractAddress']
	/** The collect modules enum */
	type: CollectModules
}
// ...
export type TimedFeeCollectModuleSettings = {
	__typename?: 'TimedFeeCollectModuleSettings'
	/** The collect module amount info */
	amount: ModuleFeeAmount
	contractAddress: Scalars['ContractAddress']
	/** The collect module end timestamp */
	endTimestamp: Scalars['DateTime']
	/** Follower only */
	followerOnly: Scalars['Boolean']
	/** The collect module recipient address */
	recipient: Scalars['EthereumAddress']
	/** The collect module referral fee */
	referralFee: Scalars['Float']
	/** The collect modules enum */
	type: CollectModules
}
// ...
export type Wallet = {
	__typename?: 'Wallet'
	address: Scalars['EthereumAddress']
	/** The default profile for the wallet for now it is just their first profile, this will be the default profile they picked soon enough */
	defaultProfile?: Maybe<Profile>
}